genium-backend/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
python-dotenv>=1.0.0
pypdf>=3.17.0 
Flask>=2.0.0
Flask-Cors>=3.0.0